+++ Common/Constants.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TheBindingOfRarria.Common;
+internal class Constants
+{
+}
