--- Content/Buffs/AnubisCurse.cs
@@ -1,21 +1,0 @@
-
-using Terraria;
-using Terraria.ModLoader;
-
-namespace TheBindingOfRarria.Content.Buffs;
-
-public class AnubisCurse : ModBuff
-{
-    public override void SetStaticDefaults()
-    {
-        Main.debuff[Type] = true;
-    }
-}
-public class MummyPlayer : ModPlayer
-{
-    public override void PostUpdate()
-    {
-        if (Player.HasBuff(ModContent.BuffType<AnubisCurse>()))
-            Player.lifeRegen -= 2;
-    }
-}
