+++ Content/Buffs/BloodShieldBleed.cs
@@ -1,0 +1,20 @@
+﻿using Terraria;
+using Terraria.ModLoader;
+using TheBindingOfRarria.Content.Items;
+
+namespace TheBindingOfRarria.Content.Buffs;
+
+public class BloodShieldBleed : ModBuff
+{
+    public override string Texture => ContentPath + "Buffs/" + Name;
+
+    public override void SetStaticDefaults() => Main.debuff[Type] = true;
+
+    public override void Update(Player player, ref int buffIndex)
+    {
+        player.lifeRegen -= player.statLifeMax2 / 10;
+        
+        player.GetModPlayer<KamoPlayer>().Stored += player.statLifeMax2 * (1f / 7200);
+        player.statDefense += (int)player.GetModPlayer<KamoPlayer>().Stored;
+    }
+}
