+++ Content/Buffs/PoisonImmunity.cs
@@ -1,0 +1,12 @@
+using Terraria;
+using Terraria.ModLoader;
+
+namespace TheBindingOfRarria.Content.Buffs;
+
+public class PoisonImmunity : ModBuff
+{
+    public override void Update(Player player, ref int buffIndex)
+    {
+        player.lifeRegen += 4;
+    }
+}
