+++ Content/Items/BrokenAnkh.cs
@@ -2,7 +2,7 @@
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
-using TheBindingOfRarria.Content.Buffs;
+using TheBindingOfRarria.Content.Buffs.Debuffs;
 
 namespace TheBindingOfRarria.Content.Items;
